refactor types into helpers file, fix type error
