+++ app/components/Editor/Sidebar/CodeEditor/index.tsx
@@ -1,30 +1,14 @@
-import axios from "axios";
 import type monacoEditor from "monaco-editor/esm/vs/editor/editor.api";
 import { useContext, useEffect, useState } from "react";
 
 import { RunnerContext } from "../../contexts/RunnerContext";
 import { TestContext } from "../../contexts/TestContext";
 import EditorComponent from "../Editor";
+import { includeTypes } from "../helpers";
 import { useEnvTypes } from "./hooks/envTypes";
 import { useGlyphs } from "./hooks/glyphs";
 
 type Editor = monacoEditor.editor.IStandaloneCodeEditor;
-
-const TYPES_URL = "/types.txt";
-
-export const includeTypes = (monaco: typeof monacoEditor) => {
-  axios.get(TYPES_URL).then(({ data: types }) => {
-    const uri = monaco.Uri.file("qawolf/types.d.ts");
-    if (monaco.editor.getModel(uri)) return;
-
-    monaco.languages.typescript.javascriptDefaults.addExtraLib(
-      types,
-      uri.toString()
-    );
-
-    monaco.editor.createModel(types, "typescript", uri);
-  });
-};
 
 export default function CodeEditor(): JSX.Element {
   // code, env, onSelectionChange,
